refactor classes out of __init__
